--- supertask/settings.py
@@ -1,15 +1,0 @@
-import dataclasses
-import typing as t
-
-from .model import JobStoreLocation
-
-
-@dataclasses.dataclass
-class Settings:
-    """
-    Bundle settings for propagating them from the environment to the FastAPI domain.
-    """
-
-    store_location: JobStoreLocation
-    pre_delete_jobs: bool
-    pre_seed_jobs: t.Optional[str]
